+++ src/types/internal.ts
@@ -16,9 +16,11 @@ export type RealAny = PromiseLike<any> | Promise<any> | any
 export type UnionToIntersection<U> =
   (U extends any ? (k: U) => void : never) extends ((k: infer I) => void) ? I : never
 
-export type ExctractParameters<Path extends string> = Path extends `${infer Segment}/${infer Rest}`
-  ? Segment extends `<${infer Param}>` ? Record<Param, string> & ExctractParameters<Rest> : ExctractParameters<Rest>
+type ExtractParametersRecord<Path extends string> = Path extends `${infer Segment}/${infer Rest}`
+  ? Segment extends `<${infer Param}>` ? Record<Param, string> & ExtractParametersRecord<Rest> : ExtractParametersRecord<Rest>
   : Path extends `<${infer Param}>` ? Record<Param, string> : {}
+
+export type ExtractParameters<Path extends string> = keyof ExtractParametersRecord<Path> extends never ? string : keyof ExtractParametersRecord<Path>
 
 export type MergeObjects<T extends object[]> = {
   [K in keyof UnionToIntersection<T[number]>]:
@@ -42,6 +44,10 @@ export type DeepRequired<Type> = Type extends Content
 	: Type extends {}
   ? { [Key in keyof Type]-?: DeepRequired<Type[Key]> }
   : Required<Type>
+
+export type DeepPartial<T> = T extends object ? {
+	[P in keyof T]?: DeepPartial<T[P]>
+} : T
 
 export type LoadPath = {
 	path: string